Last updates 7-9

+++ lib/utilits/paymob_manager.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 
 class PaymobManager {
-  Future<String> getPaymentKey(int amount, String currency) async {
+  Future<String> getPaymentKey(double amount, String currency) async {
     try {
       String authanticationToken = await _getAuthanticationToken();
